requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via
    #   datasets
    #   fsspec
aiosignal==1.2.0
    # via aiohttp
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.25.0
    # via -r requirements.in
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
black==21.12b0
    # via ipython
bleach==4.1.0
    # via nbconvert
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   argon2-cffi-bindings
    #   pynacl
charset-normalizer==2.0.10
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via black
coverage[toml]==6.3
    # via pytest-cov
cycler==0.11.0
    # via matplotlib
datasets==1.18.3
    # via -r requirements.in
debugpy==1.5.1
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
dependency-injector==4.37.0
    # via -r requirements.in
deprecated==1.2.13
    # via pygithub
dill==0.3.4
    # via
    #   datasets
    #   multiprocess
entrypoints==0.3
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.2
    # via stack-data
faker==12.3.0
    # via -r requirements.in
filelock==3.4.2
    # via huggingface-hub
fonttools==4.29.0
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2022.1.0
    # via datasets
gitdb==4.0.9
    # via gitpython
gitpython==3.1.26
    # via pydriller
huggingface-hub==0.4.0
    # via
    #   -r requirements.in
    #   datasets
idna==3.3
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
ipykernel==6.7.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.0.1
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
    #   qtconsole
ipywidgets==7.6.5
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   -r requirements.in
    #   nbconvert
    #   notebook
jsonschema==4.4.0
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.1.2
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.0
    # via jupyter
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
kiwisolver==1.3.2
    # via matplotlib
lizard==1.17.9
    # via pydriller
loguru==0.5.3
    # via -r requirements.in
markupsafe==2.0.1
    # via jinja2
matplotlib==3.5.1
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.12.2
    # via datasets
mypy-extensions==0.4.3
    # via black
nbclient==0.5.10
    # via nbconvert
nbconvert==6.4.1
    # via
    #   jupyter
    #   notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.8
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.22.1
    # via
    #   -r requirements.in
    #   datasets
    #   matplotlib
    #   pandas
    #   pyarrow
packaging==21.3
    # via
    #   bleach
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   pytest
    #   qtpy
pandas==1.4.0
    # via
    #   -r requirements.in
    #   datasets
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.0
    # via matplotlib
platformdirs==2.4.1
    # via black
pluggy==1.0.0
    # via pytest
prometheus-client==0.13.0
    # via notebook
prompt-toolkit==3.0.26
    # via
    #   ipython
    #   jupyter-console
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pyarrow==7.0.0
    # via datasets
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via -r requirements.in
pydriller==2.0
    # via -r requirements.in
pygithub==1.55
    # via -r requirements.in
pygments==2.11.2
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyjwt==2.3.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==3.0.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==0.19.2
    # via -r requirements.in
pytz==2021.3
    # via
    #   pandas
    #   pydriller
pyyaml==6.0
    # via huggingface-hub
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.2.2
    # via jupyter
qtpy==2.0.0
    # via qtconsole
requests==2.27.1
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   pygithub
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   bleach
    #   dependency-injector
    #   python-dateutil
smmap==5.0.0
    # via gitdb
stack-data==0.1.4
    # via ipython
terminado==0.13.1
    # via notebook
testpath==0.5.0
    # via nbconvert
toml==0.10.2
    # via pytest
tomli==1.2.3
    # via
    #   black
    #   coverage
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.62.3
    # via
    #   datasets
    #   huggingface-hub
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==4.0.1
    # via
    #   black
    #   huggingface-hub
    #   pydantic
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
widgetsnbextension==3.5.2
    # via ipywidgets
wrapt==1.13.3
    # via deprecated
xxhash==2.0.2
    # via datasets
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
